requirements.txt
docxtpl==0.16.7
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
Pillow==10.0.1
streamlit==1.27.2
streamlit_option_menu==0.3.6
